Update README.md
The project proposes a software to support psychiatric diagnosis that uses artificial intelligence (AI) and machine learning to help mental health professionals calculate the probability of specific diagnoses (based on ICD), combining data from neuropsychological evaluations, complementary exams, and psychiatric diagnostic criteria. The tool is intended to be a complementary aid, without replacing the clinical and subjective analysis of the professional, respecting confidentiality and the Brazilian General Data Protection Law.